szmakers/requirements.txt
Django==2.0.3
django-autoslug-iplweb==1.9.4.dev0
django-ckeditor==5.4.0
django-crispy-forms==1.7.2
django-js-asset==1.0.0
django-redis-cache==1.6.5
django-registration-redux==2.2
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
redis==2.10.5
sorl-thumbnail==12.4.1
